chore: Add FileStorage class for serializing and deserializing instances
The code changes include adding a new file called file_storage.py which contains the FileStorage class. This class is responsible for serializing instances to a JSON file and deserializing back to instances. It includes methods for retrieving all objects, adding new objects, saving objects to the JSON file, reloading objects from the JSON file, deleting objects, and closing the file storage. The class also includes methods for retrieving specific objects based on class name and ID, as well as counting the number of objects in storage.