feat: Combine product data with image URLs in get_products route
This commit modifies the get_products route in the app.py file. It combines the product data with image URLs by creating a dictionary to map product_id to image_url. The product_images_dict is then used to add the image_url to each product in the products_list. This change improves the response of the get_products route by including the image URL for each product.