Refactor requirements.txt to update dependencies and add new packages
This commit updates the requirements.txt file to include new versions of dependencies and adds new packages required for the application. The changes include updating blinker to version 1.8.2 and click to version 8.1.7. Additionally, greenlet, itsdangerous, MarkupSafe, and typing_extensions are added to the requirements. These updates and additions ensure that the application is using the latest stable versions of the required packages.